+++ src/js/services/state/state.module.js
@@ -1,0 +1,5 @@
+import service from "./state.service";
+
+angular
+.module("uploader.state", [])
+.factory("State", service);