+++ src/js/services/state/state.service.js
@@ -1,0 +1,15 @@
+state.$inject = ["$q"];
+
+export default state;
+
+function state($q) {
+  const prefix = "modwatch.";
+  return {
+    isFirstTime() {
+      return $q.resolve(!localStorage.getItem(`${prefix}notfirsttime`));
+    },
+    disableFirstTime() {
+      return $q.resolve(localStorage.setItem(`${prefix}notfirsttime`, true));;
+    }
+  };
+}