+++ src/js/services/toast/toast.service.js
@@ -5,6 +5,7 @@ import badpasswordT from "./templates/badpassword.toast.html";
 import savedinfoT from "./templates/savedinfo.toast.html";
 import versionT from "./templates/version.toast.html";
 import uploaddoneT from "./templates/uploaddone.toast.html";
+import debughelperT from "./templates/debughelper.toast.html";
 
 export default Toast;
 
@@ -17,20 +18,23 @@ function Toast($mdToast) {
       defaultToast(badpasswordT);
     },
     savedInfo() {
-      defaultToast(savedinfoT, 3);
+      defaultToast(savedinfoT, {hideDelay: 3});
     },
     uploadDone() {
       defaultToast(uploaddoneT);
     },
     version() {
       defaultToast(versionT);
+    },
+    debughelper() {
+      defaultToast(debughelperT);
     }
   };
 
-  function defaultToast(template, hideDelay) {
+  function defaultToast(template, opts = {}) {
     $mdToast.show({
       template,
-      hideDelay: hideDelay ? hideDelay * 1000 : 6000,
+      hideDelay: typeof opts.hideDelay !== "undefined" ? opts.hideDelay * 1000 : 6000,
       position: "bottom right"
     });
   }