+++ src/js/uploader.module.js
@@ -4,14 +4,14 @@ import "angular-aria";
 import "angular-material";
 
 import "./services/api/api.module";
+import "./services/state/state.module";
 import "./services/toast/toast.module";
 import "./controllers/main/main.module";
 
 angular
 .module("uploader", [
-  // "ngMaterial",
-
   "uploader.main",
   "uploader.api",
+  "uploader.state",
   "uploader.toast"
 ]);